requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# API Documentation
drf-spectacular==0.26.5

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Utilities
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.1.0